Consolidate info scripts into new code
