+++ poppunk_info-runner.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+# vim: set fileencoding=<utf-8> :
+# Copyright 2018-2020 John Lees, Daniel Anderson and Nick Croucher
+
+"""Convenience wrapper for running poppunk_info directly from source tree."""
+
+from PopPUNK.info import main
+
+if __name__ == '__main__':
+    main()
+
