+++ src/boundary.cpp
@@ -146,6 +146,31 @@ edge_tuple generate_tuples(const std::vector<int> &assignments,
     return edge_vec;
 }
 
+edge_tuple generate_all_tuples(const int num_ref,
+                               const int num_queries,
+                               bool self,
+                               const int int_offset) {
+    edge_tuple edge_vec;
+    if (self) {
+        const size_t n_rows = ((2 * num_ref - 1)**2 - 1) / 8;
+        for (long row_idx = 0; row_idx < n_rows; row_idx++) {
+            long i = calc_row_idx(row_idx, n_samples);
+            long j = calc_col_idx(row_idx, i, n_samples) + int_offset;
+            i = i + int_offset;
+            long min_node = std::min(i,j);
+            long max_node = std::max(i,j);
+            edge_vec.push_back(std::make_tuple(min_node, max_node));
+        }
+    } else {
+        for (long i = 0; i < num_ref; i++) {
+            for (long j = 0; j < num_queries; j++) {
+                edge_vec.push_back(std::make_tuple(i, j + num_ref));
+            }
+        }
+    }
+    return edge_vec;
+}
+
 // Line defined between (x0, y0) and (x1, y1)
 // Offset is distance along this line, starting at (x0, y0)
 network_coo threshold_iterate_1D(const NumpyMatrix &distMat,
